Added time/weather/location widget
Had a lot of fun building this with some help from ChatGPT- I wanted to make something I hadn't made before. The widget pulls the user's IP address and displays the time and weather at their location! I used the 'go live' plugin to test it until I got it right- the weather/location data is static but the time updates every minute. Worked on this last night in a separate file and finished/added it to the repo today.